refactor: boy i hope this works, nixing default
